push

+++ Project_CK/project_ck/src/main/java/tin/ta/project_ck/Repo/MonHocRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tin.ta.project_ck.Models.MonHocModel;
 @Repository
-public interface MonHocRepo extends JpaRepository<MonHocModel, Integer>  {
+public interface MonHocRepo extends JpaRepository<MonHocModel, Long>  {
 }
