+++ Project_CK/project_ck/src/main/java/tin/ta/project_ck/Repo/ScoreRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tin.ta.project_ck.Models.ScoreModel;
 @Repository
-public interface ScoreRepo extends JpaRepository<ScoreModel, Integer>  {
+public interface ScoreRepo extends JpaRepository<ScoreModel, Long>  {
     List<ScoreModel> findByStudentId(Long studentId);
 }
